+++ packages/shared/src/catalog/index.ts
@@ -13,6 +13,7 @@ export * from "./primitives/link";
 export * from "./primitives/numberField";
 export * from "./primitives/searchField";
 export * from "./primitives/separator";
+export * from "./primitives/slider";
 export * from "./primitives/switch";
 export * from "./primitives/textField";
 export * from "./primitives/timeField";